+++ src/hooks/locals/ru.ts
@@ -1,0 +1,27 @@
+import {Translations} from "@/hooks/locals/locales.ts";
+
+const ru: Translations = {
+    loginPage: {
+        heading: "Вход на сайт",
+        submit: "Войти",
+        login: {
+            label: "Логин или Email",
+            placeholder: "Введите ваш логин или Email",
+            required: "Это поле обязательно для заполнения",
+            invalid: "Неверный формат",
+            minLength: "Минимальная длина: 1 символ",
+            maxLength: "Максимальная длина: 255 символов"
+        },
+        password: {
+            label: "Пароль",
+            placeholder: "Введите ваш пароль",
+            required: "Это поле обязательно для заполнения",
+            invalid: "Неверный пароль",
+            minLength: "Минимальная длина: 8 символов",
+            maxLength: "Максимальная длина: 255 символов"
+        },
+        accountNotExists: "Отсутсвует аккаунт?"
+    }
+}
+
+export default ru;
